+++ src/solana/SolanaEngine.ts
@@ -1,6 +1,6 @@
 import {
   createAssociatedTokenAccountInstruction,
-  createTransferInstruction,
+  createTransferCheckedInstruction,
   getAssociatedTokenAddressSync
 } from '@solana/spl-token'
 import {
@@ -165,19 +165,33 @@ export class SolanaEngine extends CurrencyEngine<
             { programId: this.networkInfo.tokenPublicKey },
             { encoding: 'jsonParsed' }
           ]
+        },
+        {
+          method: 'getTokenAccountsByOwner',
+          params: [
+            this.base58PublicKey,
+            { programId: this.networkInfo.token2022PublicKey },
+            { encoding: 'jsonParsed' }
+          ]
         }
       ]
 
       const balances: any = await this.fetchRpcBulk(requests)
 
-      const [mainnetBal, tokenBalsRaw]: [AccountBalance, TokenAmount[]] =
-        balances
+      const [mainnetBal, tokenBalsRaw, token2022BalsRaw]: [
+        AccountBalance,
+        TokenAmount[],
+        TokenAmount[]
+      ] = balances
 
       const balance = asAccountBalance(mainnetBal)
       this.updateBalance(this.chainCode, balance.result.value.toString())
       this.usedTokenIdSet.add(null)
 
-      const tokenBalances = asTokenBalances(tokenBalsRaw).result.value
+      const tokenBalances = [
+        ...asTokenBalances(tokenBalsRaw).result.value,
+        ...asTokenBalances(token2022BalsRaw).result.value
+      ].flat()
       const tokenBalMap = tokenBalances.reduce(
         (acc: { [key: string]: string }, tokenBal) => {
           acc[tokenBal.account.data.parsed.info.mint] =
@@ -335,8 +349,10 @@ export class SolanaEngine extends CurrencyEngine<
     const preTokenBalances = maybePreTokenBalances ?? []
     const postTokenBalances = maybePostTokenBalancesRaw ?? []
     const isTokenTransaction =
-      tx.transaction.message.staticAccountKeys.find(pk =>
-        pk.equals(this.tools.tokenProgramPublicKey)
+      tx.transaction.message.staticAccountKeys.find(
+        pk =>
+          pk.equals(this.tools.tokenProgramPublicKey) ||
+          pk.equals(this.tools.token2022ProgramPublicKey)
       ) != null
     const networkFee = fee.toString()
 
@@ -408,19 +424,21 @@ export class SolanaEngine extends CurrencyEngine<
     this.updateTxStatus(this.currencyInfo.currencyCode, 1)
 
     for (const tokenId of this.enabledTokenIds) {
-      const cc = this.allTokensMap[tokenId].currencyCode
+      const token = this.allTokensMap[tokenId]
       if (this.usedTokenIdSet.has(tokenId)) {
+        const tokenOwnerPubkey = this.tools.getTokenOwnerPublicKey(token)
+
         const pk = getAssociatedTokenAddressSync(
           new PublicKey(tokenId),
           mainPubkey,
           false,
-          this.tools.tokenProgramPublicKey,
+          tokenOwnerPubkey,
           new PublicKey(this.networkInfo.associatedTokenPublicKey)
         )
-        await this.queryTransactionsInner(cc, pk)
+        await this.queryTransactionsInner(token.currencyCode, pk)
         this.sendTransactionEvents()
       }
-      this.updateTxStatus(cc, 1)
+      this.updateTxStatus(token.currencyCode, 1)
     }
   }
 
@@ -669,7 +687,11 @@ export class SolanaEngine extends CurrencyEngine<
 
     if (isTokenTx) {
       const tokenPubkey = new PublicKey(tokenId)
-      const tokenProgramId = this.tools.tokenProgramPublicKey
+      const token = this.allTokensMap[tokenId]
+      const tokenOwnerPubkey = this.tools.getTokenOwnerPublicKey(token)
+
+      const decimals =
+        token.denominations[0].multiplier.match(/0/g)?.length ?? 0
       const associatedTokenProgramId = new PublicKey(
         this.networkInfo.associatedTokenPublicKey
       )
@@ -679,7 +701,7 @@ export class SolanaEngine extends CurrencyEngine<
         tokenPubkey,
         payee,
         true, // payee may be a Program Derived Address
-        tokenProgramId,
+        tokenOwnerPubkey,
         associatedTokenProgramId
       )
 
@@ -712,7 +734,7 @@ export class SolanaEngine extends CurrencyEngine<
             associatedDestinationTokenAddrPayee,
             payee,
             tokenPubkey,
-            tokenProgramId,
+            tokenOwnerPubkey,
             associatedTokenProgramId
           )
         instructions.push(tokenCreationInstruction)
@@ -740,16 +762,18 @@ export class SolanaEngine extends CurrencyEngine<
         tokenPubkey,
         payer,
         false,
-        tokenProgramId,
+        tokenOwnerPubkey,
         associatedTokenProgramId
       )
-      const tokenTransferInstruction = createTransferInstruction(
+      const tokenTransferInstruction = createTransferCheckedInstruction(
         associatedDestinationTokenAddrPayer,
+        new PublicKey(tokenId),
         associatedDestinationTokenAddrPayee,
         payer,
         BigInt(nativeAmount),
+        decimals,
         [],
-        tokenProgramId
+        tokenOwnerPubkey
       )
       instructions.push(tokenTransferInstruction)
     } else {